requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML>=6.0
openpyxl>=3.1.5
pandas>=2.1.0
sqlalchemy==2.0.34
psycopg[binary]==3.2.1
alembic==1.13.2
tenacity==9.0.0
jsonschema==4.23.0
pymysql==1.1.1
flask==3.1.2
